Upload second copy of ODP and OSL for inclusion GP5 releases
The PXF release process currently uploads ODP and OSL artifacts to a
GP6-specific location in blob store and directs the RelEng team to use
the same artifacts for PXF-GP5 releases. The RelEng team would prefer
that the ODP and OSL artifacts get uploaded to GP5 and GP6 locations.

Instead of adding additional template variables for GP5 and GP6 paths
for ODP and OSL, the existing template variables are updated to remove
the GP major version from the path. This is appended in the CI pipeline
instead. Since all release artifacts for RelEng are stored in the same
bucket, the template variables for the bucket name are simplified as
well.

PXF ODP and OSL artifacts for both GP5 and GP6 are now stored in Google
Cloud Storage so there isn't much benefit to parameterizing the protocol
(i.e., `s3` vs `gs`).

This commit allows for using a common set of template variables files for the
build pipelines on master and branch-5.16.

Co-authored-by: Ashuka Xue <axue@vmware.com>
Co-authored-by: Bradford D. Boyle <bradfordb@vmware.com>